requirements.txt
requests
streamlit
langchain_aws
langchain_core
langchain
langchain_community
python-dotenv
geopy
scikit-learn
openpyxl
pymysql
psycopg2